autonomica-api/requirements.txt
# Core FastAPI Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3

# AI and LLM Dependencies
openai==1.3.5
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10

# Agent Framework Dependencies
# Note: OWL and CAMEL frameworks would be installed from source or custom repositories
# For now, we'll use compatible agent libraries
autogen-agentchat==0.2.0
multi-agent-orchestrator==0.1.2

# Database and Storage
redis==5.0.1
faiss-cpu==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1

# Vector Storage and Embeddings
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.1.3

# Background Tasks and Queue
celery==5.3.4
flower==2.0.1

# HTTP and WebSocket Support
httpx==0.25.2
websockets==12.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring and Logging
python-dotenv==1.0.0
loguru==0.7.2
prometheus-client==0.19.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Data Validation and Serialization
marshmallow==3.20.1
orjson==3.9.10

# Async HTTP Client
aiohttp==3.9.1
aiofiles==23.2.1

# Configuration Management
dynaconf==3.2.4 